Fixed a silly typo in ai_requests.py's __main__ function for testing Ollama models in the future.
Improved fallback for the 1/1000 chance two annotators submit at the same time and the system attempts to give them both the same paper. During testing the slower submitter would get locked out (Appropriately) but the system would fail to attempt to search for another paper, causing the user to refresh the webpage.